\name{MAGEE-package}
\alias{MAGEE-package}
\docType{package}
\title{
Mixed Model Association Test for GEne-Environment Interaction
}
\description{
An R package for performing generalized linear mixed model-based single-variant and variant set main effect tests, gene-environment interaction tests, and joint tests for association, as proposed in Wang et al. (2020) <DOI:10.1002/gepi.22351>.
}
\details{
\tabular{ll}{
Package: \tab MAGEE\cr
Type: \tab Package\cr
Version: \tab 1.4.5\cr
Date: \tab 2026-02-11\cr
License: \tab GPL (>= 3)\cr
}
}
\author{
Xinyu Wang, Han Chen, Duy T. Pham

Maintainer: Han Chen <Han.Chen.2@uth.tmc.edu>
}
\references{
Wang, X., Lim, E., Liu, C, Sung, Y.J., Rao, D.C., Morrison, A.C., Boerwinkle, E., Manning, A. K., and Chen, H. (2020) Efficient gene-environment interaction tests for large biobank-scale sequencing studies. Genetic Epidemiology, 44(8): 908-923.

Chen, H., Huffman, J.E., Brody, J.A., Wang, C., Lee, S., Li, Z., Gogarten, S.M., Sofer, T., Bielak, L.F., Bis, J.C., et al. (2019) Efficient variant set mixed model association tests for continuous and binary traits in large-scale whole-genome sequencing studies. The American Journal of Human Genetics, 104 (2): 260-274.

Wang, X., Pham, D.T., Westerman, K.E., Pan, C., Manning, A.K., Chen, H., (2022) Genomic summary statistics and meta-analysis for set-based gene-environment interaction tests in large-scale sequencing studies. medrxiv. 
}
\keyword{ package }
\keyword{ generalized linear mixed model }
\keyword{ gene-environment interaction }
\keyword{ joint test }
